requirements.txt
amqp==5.0.9
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==37.0.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.9
django-allauth==0.50.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-heroku==0.3.1
django-redis==5.2.0
django-rq==2.5.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
kombu==5.2.3
oauthlib==3.2.0
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pymemcache==3.5.0
pyparsing==3.0.6
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==4.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rq==1.10.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
